Gun-3-Ders-1/iristahmini/requirements.txt
streamlit
pandas
matplotlib
numpy
scikit-learn
seaborn
Pillow

